refactor(web): extract backtesting export handlers
